+++ pso/objectivefunction.h
@@ -10,18 +10,18 @@ class ObjectiveFunction
 {
 public:
     ObjectiveFunction(ILog &logger,
-                      double (*objectiveFunction)(std::vector<int>&),
+                      double (*objectiveFunction)(std::vector<double>&),
                       int variablesCount);
 
-    double GetResult(std::vector<int>& input);
+    double GetResult(std::vector<double>& input);
     int GetVariablesCount();
 
 private:
-    double (*_objectiveFunction)(std::vector<int>&);
+    double (*_objectiveFunction)(std::vector<double>&);
     ILog *_logger;
     int _variablesCount;
 
-    void ValidateInputSize(std::vector<int>& input);
+    void ValidateInputSize(std::vector<double>& input);
 };
 
 #endif // OBJECTIVEFUNCTION_H
